✨  Added logic for saving weapon locally ✨

+++ data/network/src/main/kotlin/com/example/valorant/data/network/model/dto/weapon/detail/stats/DamageRangesDTO.kt
@@ -12,7 +12,7 @@ data class DamageRangesDTO(
     @SerialName("headDamage")
     val headDamage: Double,
     @SerialName("bodyDamage")
-    val bodyDamage: Int,
+    val bodyDamage: Double,
     @SerialName("legDamage")
     val legDamage: Double,
 )
